s4:dsdb: Use appropriate type for userAccountControl flags

+++ source4/dsdb/common/util.c
@@ -645,7 +645,7 @@ NTSTATUS samdb_result_passwords(TALLOC_CTX *mem_ctx,
 				const struct ldb_message *msg,
 				struct samr_Password **nt_pwd)
 {
-	uint16_t acct_flags;
+	uint32_t acct_flags;
 
 	acct_flags = samdb_result_acct_flags(msg,
 					     "msDS-User-Account-Control-Computed");
